Cleanup up warnings & stuff
git-svn-id: http://svn.code.sf.net/p/plarpebu/code@51 08b1f3f0-860d-0410-8033-80c2cf11100b

+++ src/plugins/karaoke/cdg/io/CdgFileObject.java
@@ -35,7 +35,7 @@ public class CdgFileObject {
 
         // Allocate the array that will contain all the CdgDataChunks
         int fileLength = (int) file.length();
-        int nbOfCdgDataChunks = (int) (fileLength / 24);
+        int nbOfCdgDataChunks = fileLength / 24;
         // CdgDataChunk []cdgDataChunksArrayTmp = new
         // CdgDataChunk[nbOfCdgDataChunks];
         cdgDataChunksArray = new CdgDataChunk[nbOfCdgDataChunks];
